Move callNodePathForStack to profile-data.js
